Move Texture storage to separate files
