+++ Tests/Renderer/Textures/TextureStorage.cpp
@@ -1,6 +1,7 @@
 #include <gtest/gtest.h>
 
 #include <Renderer/Textures/TextureStorage.h>
+#include <Renderer/Textures/TextureLinearStorage.h>
 
 #include <Renderer/Colours.h>
